requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-filter>=24.0
django-cors-headers>=4.4
drf-spectacular>=0.28
psycopg[binary]>=3.2
python-decouple>=3.8
requests>=2.32
